requirements.txt
# =============================================================================
# Research Paper RAG Assistant — requirements.txt
# Python 3.11+  |  pip install -r requirements.txt
# =============================================================================

# --- PDF parsing ---
pymupdf>=1.24.0

# --- RAG orchestration ---
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0


# --- Vector store ---
chromadb>=0.5.0
pinecone>=5.0.0

# --- LLM and embedding providers ---
openai>=1.50.0
anthropic>=0.39.0
tiktoken>=0.8.0
litellm>=1.0.0
groq>=0.9.0

# --- Reranking ---
cohere>=5.0.0
sentence-transformers>=3.0.0

# --- Evaluation ---
ragas>=0.2.0,<0.3.0
datasets>=2.18.0

# --- Experiment tracking ---
mlflow>=2.16.0

# --- UI ---
streamlit>=1.40.0

# --- Config and validation ---
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# --- Resilience ---
tenacity>=9.0.0
httpx>=0.27.0

# --- Logging ---
structlog>=24.0.0


# --- Dev helpers ---
rich>=13.0.0

plotly